Refactored code and set up states to use
