+++ src/App.js
@@ -1,17 +1,14 @@
 import React from 'react'
-import Header from './Header';
-import Inputbox from './Inputbox';
+import Header from './components/Header';
+import ZipInput from "./components/ZipInput";
 
 
 export default function App() {
   return (
-    <div>
-
-      <Header text="Zip Code Search" color="red" />
-      <Inputbox />
-
-
-    </div>
+      <div>
+          <Header text="Zip Code Search"></Header>
+          <ZipInput></ZipInput>
+      </div>
   );
 }
