--- src/Inputbox.js
@@ -1,19 +1,0 @@
-import React from "react";
-
-function Inputbox(){
-
-   return ( 
-   
-   <div className="box" >
-
-       <label id="label">Zip Code:</label> 
-
-       <input type="text" name="lname" />  <br /> <br />
-
-       <label id="label">No Result</label>
-
-   </div>)
-
-
-}
-export default Inputbox;
