+++ src/stylesheet.css
@@ -6,6 +6,8 @@
 
 }
 
+
+
 #text{
 
     padding-top: 100px;
@@ -22,6 +24,7 @@
     padding-left: 900px;
 }
 
-.label{
+#label{
     font-weight: bold;
+    margin-right: 5px;
 }
